requirements.txt
# Core dependencies for the email application
python-dotenv==1.0.0
